+++ levior/__main__.py
@@ -10,10 +10,14 @@ from typing import Union, TextIO
 from aiogemini.tofu import create_server_ssl_context
 from aiogemini.server import Server
 
+import appdirs
+import diskcache
+
 from omegaconf import OmegaConf
 from omegaconf import DictConfig
 from omegaconf import ListConfig
 
+from . import __appname__
 from . import default_cert_paths
 from . import ocresolvers  # noqa
 from .handler import create_levior_handler
@@ -125,7 +129,9 @@ def get_config(args) -> DictConfig:
         'port': args.port,
         'cache_path': args.cache_path,
         'cache_enable': args.cache_enable,
-        'cache_ttl_default': args.cache_ttl_secs,
+        'cache_ttl_default': args.cache_ttl_default,
+        'cache_size_limit': args.cache_size_limit,
+        'cache_eviction_policy': args.cache_eviction_policy,
         'verify_ssl': args.verify_ssl,
         'socks5_proxy': args.socks5_proxy,
         'tor': args.tor,
@@ -138,7 +144,7 @@ def get_config(args) -> DictConfig:
         'urules': [{
             'regex': r'.*',
             'cache': args.cache_enable,
-            'ttl': args.cache_ttl_secs
+            'ttl': args.cache_ttl_default
         }]
     })
 
@@ -159,6 +165,36 @@ def get_config(args) -> DictConfig:
     return config, rules
 
 
+def configure_cache(config: DictConfig) -> diskcache.Cache:
+    cache_dir: Path = Path(
+        config.cache_path if config.cache_path else
+        appdirs.user_cache_dir(__appname__)
+    )
+
+    if not cache_dir.exists():
+        cache_dir.mkdir(parents=True)
+
+    if config.cache_eviction_policy in ['least-recently-stored',
+                                        'least-recently-used',
+                                        'least-frequently-used',
+                                        'none']:
+        cpolicy = config.cache_eviction_policy
+    else:
+        cpolicy = 'least-recently-stored'
+
+    if isinstance(config.cache_size_limit, (int, float)) and \
+       config.cache_size_limit > 0:
+        size_limit_mb = config.cache_size_limit
+    else:
+        size_limit_mb = 2048
+
+    return diskcache.Cache(
+        str(cache_dir),
+        eviction_policy=cpolicy,
+        size_limit=size_limit_mb * 1024 * 1024
+    )
+
+
 def levior_configure_server(args) -> (DictConfig, Server):
     """
     Create a levior server from the command-line config arguments
@@ -169,6 +205,7 @@ def levior_configure_server(args) -> (DictConfig, Server):
 
     config, rules = get_config(args)
 
+    data_dir: Path = Path(appdirs.user_data_dir(__appname__))  # noqa
     try:
         for mode in config.mode.split(','):
             if not mode:
@@ -185,7 +222,7 @@ def levior_configure_server(args) -> (DictConfig, Server):
 
     return (config, Server(
         create_server_ssl_context(cert_path, key_path),
-        create_levior_handler(config, rules),
+        create_levior_handler(config, configure_cache(config), rules),
         host=config.get('hostname', 'localhost'),
         port=config.get('port', 1965)
     ))